Fix rebind issue in ServiceWatcher
ServiceWatcher is not rebinding properly when the binder dies. This
appears to have been a very long standing bug - it's not entirely clear
how this was never seen before. Possibly package changes after upgrade
always forced a rebind anyways. When gcore is installed by the
LocationHostTest it breaks ServiceWatcher's binding to gcore, and this
is never refreshed until there are further gcore package changes. This
was possibly exacerbated by a race condition in LocationProviderProxy
which could make the provider appear enabled even though it wasn't.

Bug: 148242856
Test: atest GtsGmscoreHostTestCases:LocationHostTest#testLocationProviders
Change-Id: Ibc11e5b0f04e16dc2d3734e02b271358c884072e