Controls UI - ActivityView WIP
Beginning elements of using an ActivityView in lieu of detail
panels. The concept works, but more refinement is needed. We would
want to limit to a single activity and properly handle events to close
the dialog. To test, set this flag:

adb shell settings put secure systemui.controls_use_detail_panel 1

Bug: 148207527
Test: manual
Change-Id: Id8a213f949a6a759c9a014c05ffa1135e800b4fc