Remove redundant mFocusedApp from DisplayPolicy
We don't really use mFocusedApp in DisplayPolicy, but just print it out.
It may confuse the reader because we already have one in DisplayContent.
And these mFocusedApp fields can reference to different apps, because
the one in DisplayPolicy will stop updating while the DisplayPolicy
doesn't have a focused window.

Fix: 141869496
Test: Manual and presubmit
Change-Id: Ic6edbade2af24135ee86d4bd9b6d8965e61fa664