Don't kill a new process for the sins of an old process.
I have a test which instruments com.android.vending, which contains
some content providers.  Each test case starts the app and then fairly
quickly kills it. I doubt it ever gets to the stage where the content
providers are working.

When the test runs slowly enough, a new instance of the app gets
killed 10s after an old instance was started, which breaks
things. That seems like the wrong behavior in general, although it
should be rare.

This change attempts to prevent that happening by cleaning up any old
timeout messages for the app when we restart it.

Bug: 149150665
Test: atest FrameworksCoreTests:android.content.ContentResolverTest
Change-Id: I1b04eab6b0fd3d24c0a2ee315818d58de4c5915d