Prevent early destroying HotspotTile
Due to the switch to Async HotspotControllerImpl, if the controller is
created with the tile it will think that tethering is not available
(because it hasn't received callbacks saying otherwise yet) and the tile
will be preemptively destroyed. This was observable on OTAs (or when
doing a full flash).

With this change, HotspotControllerImpl will assume that tethering is
available by default and will signal callbacks if it learns that it's
not.

There are currently no other consumers of
HotspotControllerImpl#isHotspotSupported.

Fixes: 149492796
Test: atest HotspotControllerImplTest
Test: manual full flash
Change-Id: Id1461b7fbe973b732bc69c7ad77a30e0f66247d8