Add key layouts for Xbox clones
There are a lot of controllers that are essentially 'clones' of the xbox
controllers. They follow the same key layouts.

To support these clones, we add these layouts as a copy of the xbox
layout. They were generated using a small python script.

The layouts will be manually verified for now, on a select subset of
these controllers.

Test: these will be selectively verified manually
Test: we will add CTS for some of these in the future

Bug: 144469012
Bug: 136263708
Change-Id: Id453f103824a1ba2bf03a9309378dafde418e594