Expose Bitmap#getHardwareBuffer
Bug: 148155907
Bug: 150395371
Test: I652101b6a4b3d9de4e77a316ba0436ce466f3620

Bitmap already has a method to create a Bitmap from a HardwareBuffer,
and the NDK allows doing the reverse. This updates the SDK to allow
retrieving the HardwareBuffer, too.

Update the docs to match wrapHardwareBuffer regarding modifications.

Throw an Exception if called on a non-HARDWARE Bitmap, which is
friendlier than the crash that previously happened on the hidden API
(and the one it replaced).

Change-Id: I408cff708efa76f89e0c4c6dae16f19166ffe74d