Bind children dynamically in new pipeline (2/2)
Support setting a logical child count on the notification group view.
Since we are filtering out the notifications and their views entirely in
the new pipeline, we can't just use the number of children views on the
notification group view to determine the "+X" value in the group view UX
for how many children are in the unexanded group. So instead we set this
value directly.

Bug: 145748993
Test: add notification group and see overflow values are accurate in
both old and new pipeline

Change-Id: I60bc58994a77bc3801f062dedc41faaee1d48494