Call #clearSettingsProvider in PowerManagerServiceTest
Add calls to FakeSettingsProvider.clearSettingsProvider in setUp and
tearDown.

While PowerManagerServiceTest uses FakeSettingsProvider to
interact with a fake version of Settings and stay independent of system
state, this does not work if run together with other tests that do
access the real Settings, as Settings caches the first ContentProvider
it sees in the process.

Test: atest com.android.server.power.PowerManagerServiceTest
Bug: 151114486
Change-Id: I04d11cfcaef89dc1832a5f57af21c815e7861d11