Prevent waiting when inject event without animation
Synchronously inject input event would wait for animations complete,
but if no surface animation is on going and 'scheduleAnimation' is
called in order to apply the pending transaction. It would wait
until timeout(=5s) because no 'notify' would be called from
framecallback. That would cause timeout problem when emulate swipe
gesture via inject many events.

This would provide a flag in WindowAnimator to ensure the pending
transaction could be applied and exit waiting if there is no
other animation.

Bug: 150250453
Bug: 152478735
Bug: 152462354
Test: atest libinput_tests inputflinger_tests
Test: atest --rerun-until-failure 100 WindowFocusTests#testMovingDisplayToTopByKeyEvent
Test: atest MemoryTests#testActivityRecreation
Test: Tests have inject events.

Change-Id: If1db5c4d923ddbcfcd6eb3db38a41fee70913568